+++ lib/langchain/llm/response/anthropic_response.rb
@@ -10,6 +10,10 @@ module Langchain::LLM
       completions.first
     end
 
+    def chat_completion
+      raw_response.dig("content", 0, "text")
+    end
+
     def completions
       [raw_response.dig("completion")]
     end
@@ -26,8 +30,20 @@ module Langchain::LLM
       raw_response.dig("log_id")
     end
 
+    def prompt_tokens
+      raw_response.dig("usage", "input_tokens").to_i
+    end
+
+    def completion_tokens
+      raw_response.dig("usage", "output_tokens").to_i
+    end
+
+    def total_tokens
+      prompt_tokens + completion_tokens
+    end
+
     def role
-      "assistant"
+      raw_response.dig("role")
     end
   end
 end